+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## Unpublished
 - Use `file-type` to determine file extension when missing
 - Remove multiple spaces in title
+- Remove default exports on sites
+- Fix Wandering Inn valid chapter check to detect patreon only chapters
 
 ## 2024-02-13 v0.1.0
 - Update `@dylanarmstrong/nodepub` version